requirements.txt
Django==5.0.2
cryptography==42.0.2
face-recognition==1.3.0
Pillow==10.2.0
python-dotenv==1.0.1
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
whitenoise==6.6.0
gunicorn==21.2.0 